requirements.txt
absl-py==2.1.0
accelerate==0.25.0
addict==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioprometheus==23.12.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Automat==20.2.0
Babel==2.15.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
bleach==6.1.0
blinker==1.9.0
blis==0.7.11
bs4==0.0.2
build==0.10.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
cloud-init==23.4.4
cloudpathlib==0.19.0
cloudpickle==3.0.0
cmake==3.31.0.1
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
Counter==1.0.0
crcmod==1.7
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.8
Cython==0.29.33
dataclasses==0.6
datasets==2.15.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
deepspeed==0.15.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
distro-info==1.1+ubuntu0.2
easyocr==1.7.2
editdistance==0.8.1
einops==0.7.0
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.1
executing==2.0.1
Faker==18.9.1
fastapi==0.114.1
faster-coco-eval==1.6.5
fastjsonschema==2.20.0
filelock==3.16.1
flash-attn==2.3.6
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.6.0
gensim==4.3.3
gguf==0.9.1
ghp-import==2.1.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.26.5
hyperlink==21.0.0
idna==3.10
imageio==2.36.0
img2pdf==0.5.1
img2table==1.4.0
importlib_metadata==8.5.0
incremental==21.3.0
iniconfig==2.0.0
inquirerpy==0.3.4
interegular==0.3.3
iotop==0.6
ipykernel==6.29.4
ipython==8.25.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.5.0
JLDracula==0.2.0
jmespath==0.10.0
joblib==1.4.2
JPype1==1.5.0
json5==0.9.25
jsonlines==4.0.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
julia==0.6.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kagglehub==0.3.3
karateclub==1.3.3
keras==3.6.0
keyring==23.5.0
kiwisolver==1.4.5
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3#subdirectory=kobert_hf
kobert-transformers==0.5.1
konlpy==0.6.0
langcodes==3.4.0
language_data==1.2.0
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
Levenshtein==0.26.1
libclang==18.1.1
lightning-utilities==0.11.7
lit==18.1.8
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistral_common==1.4.0
mistune==3.0.2
mkdocs==1.4.3
ml-dtypes==0.4.1
mmcv-full==1.7.2
mmdet==3.3.0
mmengine==0.10.5
model-index==0.1.11
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.10
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
node2vec==0.5.0
notebook_shim==0.2.4
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai==1.45.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.1
opt_einsum==3.4.0
optree==0.13.0
ordered-set==4.1.0
orjson==3.10.7
oss2==2.17.0
outlines==0.0.46
overrides==7.7.0
packaging==23.1
pandas==1.3.5
pandas-stubs==2.0.2.230605
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pathlib==1.0.1
pdf2image==1.15.1
pdf2jpg==1.1
peft==0.7.1
pexpect==4.8.0
pfzy==0.3.4
pikepdf==9.4.0
Pillow==9.5.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
polars==1.14.0
portalocker==2.10.1
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.46
protobuf==5.28.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyGObject==3.42.1
PyGSP==0.5.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.2.0
pyrsistent==0.18.1
pyserial==3.5
pytesseract==0.3.10
pytest==8.3.3
python-apt==2.4.0+ubuntu3
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.26.1
python-louvain==0.16
python-magic==0.4.24
pytorch-crf==0.7.2
pytorch-lightning==2.4.0
pytz==2023.4
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.0.3
quantile-python==1.1
RapidFuzz==3.11.0
ray==2.35.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rouge_score==0.1.2
rpds-py==0.19.1
rsa==4.9
sacrebleu==2.4.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.8.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
service-identity==18.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smart-open==7.0.4
sniffio==1.3.1
sos==4.5.6
soupsieve==2.5
spacy==3.8.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.38.5
svgwrite==1.4.3
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.7.0
timm==0.9.16
tinycss2==1.3.0
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.1
torch-geometric==2.6.1
torchaudio==2.5.1+cu121
TorchCRF==1.1.0
torchmetrics==1.4.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.29.2
trash-cli==0.17.1.14
triton==3.1.0
Twisted==22.1.0
typer==0.12.5
types-python-dateutil==2.9.0.20240316
types-pytz==2024.2.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
vllm-flash-attn==2.6.1
wadllib==1.3.6
wasabi==1.1.3
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.1.3
wrapt==1.16.0
xkit==0.0.0
XlsxWriter==3.2.0
xxhash==3.5.0
yapf==0.40.2
yarl==1.11.1
zipp==3.20.2
zope.interface==5.4.0
